CF2-012: Refactor utility imports and enhance component structure
- Updated import paths for utility functions to use the new naming convention (e.g., cn, handleRequest, toBRLCurrency).
- Removed unused Command component and its related hooks.
- Consolidated balance chart related files and updated their imports.
- Refactored useLogout hook to a new file structure.
- Enhanced the BalanceChart component with improved schema and utility functions.
- Cleaned up various components to ensure consistent utility usage across the codebase.
