+++ src/components/AlertDialog/AlertDialog.component.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as AlertDialogPrimitive from "@radix-ui/react-alert-dialog";
 
-import { cn } from "@/utils/cn";
+import { cn } from "@/utils/cn.utils";
 import { buttonVariants } from "@/components/Button/Button.component";
 
 function AlertDialog({
