+++ src/components/Card/Card.component.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import { cn } from "@/utils/cn";
+import { cn } from "@/utils/cn.utils";
 
 function Card({ className, ...props }: React.ComponentProps<"div">) {
   return (
