+++ src/components/Command/Command.component.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { Command as CommandPrimitive } from "cmdk";
 import { SearchIcon } from "lucide-react";
 
-import { cn } from "@/utils/cn";
+import { cn } from "@/utils/cn.utils";
 import {
   Dialog,
   DialogContent,
