+++ src/components/Input/Input.component.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import { cn } from "@/utils/cn";
+import { cn } from "@/utils/cn.utils";
 
 interface InputProps extends React.ComponentProps<"input"> {
   icon?: React.ReactNode;
