+++ src/components/Label/Label.component.tsx
@@ -3,7 +3,7 @@
 import * as React from "react";
 import * as LabelPrimitive from "@radix-ui/react-label";
 
-import { cn } from "@/utils/cn";
+import { cn } from "@/utils/cn.utils";
 
 function Label({
   className,
