+++ src/components/LoadingSpinner/LoadingSpinner.component.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/utils/cn";
+import { cn } from "@/utils/cn.utils";
 
 interface LoadingSpinnerProps {
   size?: "sm" | "md" | "lg";
