+++ src/components/Select/Select.component.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import * as SelectPrimitive from "@radix-ui/react-select";
 import { CheckIcon, ChevronDownIcon, ChevronUpIcon } from "lucide-react";
 
-import { cn } from "@/utils/cn";
+import { cn } from "@/utils/cn.utils";
 
 function Select({
   ...props
