+++ src/components/Table/Table.component.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import { cn } from "@/utils/cn";
+import { cn } from "@/utils/cn.utils";
 
 function Table({ className, ...props }: React.ComponentProps<"table">) {
   return (
