+++ src/components/Textarea/Textarea.component.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import { cn } from "@/utils/cn";
+import { cn } from "@/utils/cn.utils";
 
 function Textarea({ className, ...props }: React.ComponentProps<"textarea">) {
   return (
