+++ src/screens/Home/components/BalanceChart/utils/handleInitialDate.utils.ts
