+++ src/screens/Home/components/Header/Header.component.tsx
@@ -6,7 +6,7 @@ import {
   CardDescription,
   CardAction
 } from "@/components/Card/Card.component";
-import { useLogout } from "@/hooks/useLogout";
+import { useLogout } from "@/hooks/useLogout.hook";
 import { useAppSelector } from "@/store";
 import { User, LogOut } from "lucide-react";
