+++ src/screens/Register/components/RegisterForm/RegisterForm.component.tsx
@@ -28,7 +28,7 @@ import {
 import { useState } from "react";
 import { useNavigate } from "react-router";
 import { useRegisterMutation } from "@/store/services/auth/auth.service";
-import { handleRequest } from "@/utils/handleRequest";
+import { handleRequest } from "@/utils/handleRequest.utils";
 import { toast } from "sonner";
 import { useAppDispatch } from "@/store";
 import { AuthActions } from "@/store/slices/auth/auth.slice";
